requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
plyer>=2.1.0
pdfplumber>=0.10.3
openai>=1.12.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.41.0
reportlab>=4.0.0
